requirements.txt
MySQL-python
PyYAML
beanstalkc
protobuf
bernhard
boto
configobj
docker-py
kitchen
mock==1.0.1
psutil
pymongo<3.0
# Disable due to bad test cases
# https://github.com/BrightcoveOS/Diamond/issues/650
#pyrabbit
python-statsd
pyutmp
redis
simplejson
setproctitle
psycopg2
PySensors
